+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
+## Added
+- A new transformer is available under the `minecraft` name. Mapped messages will be in the form of minecraft JSON.
+Use this with `/tellraw @a` for a quick and easy minecraft bridge that supports full colours, formatting, and URLs
 ## Fixed
 - Game stdio is now run though transformers before they are processed
 - IRC no longer quits instantly after attempting to reconnect due to lag issues